renamed files, new utils, decomposed much
