+++ cogs/error_handler.py
